+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # ChangeLog
 
-## 0.7.1 Under development
+## 0.7.1 February 15, 2026
 
 - Bug #50: Serialize boolean values in `aria-*`, `data-*`, `data-ng`, `ng-*`, and `on*` attributes as explicit strings (@terabytesoftw)
 - Bug #51: Support closure values in attribute handling and add test case for boolean closure values (@terabytesoftw)